Add new base class TeamAwareTrackingAgent.
This base agent can be used if the agent should know it's current Team and provides some team related utility functions
